fix toast under notch; refactor toast
