+++ src/index.tsx
@@ -20,7 +20,7 @@ import { LnurlProvider } from './providers/lnurl'
 import { beforeSend, scrubBreadcrumb, shouldInitializeSentry } from './lib/sentry'
 import { FeesProvider } from './providers/fees'
 import { AnnouncementProvider } from './providers/announcements'
-import { ToastProvider } from './components/Toast'
+import { ToastProvider } from './providers/toast'
 import ErrorBoundary from './components/ErrorBoundary'
 import { DevModeProvider } from './providers/devMode'
 import { AssetsProvider } from './providers/assets'
